Cleaning up error handle for server to be in routes folder
